Replace context with Sylius standard and adjust ManagingProductsContext
