+++ tests/Behat/Resources/services.xml
@@ -19,8 +19,6 @@
             <argument type="service" id="sylius.repository.product_translation" />
             <argument type="service" id="sylius.context.locale" />
             <argument type="service" id="sylius.generator.slug" />
-            <argument type="service" id="sylius.behat.java_script_test_helper" />
-            <argument type="service" id="sylius.behat.notification_checker" />
         </service>
 
         <service id="webgriffe_sylius_akeneo.behat.context.setup.akeneo" class="Tests\Webgriffe\SyliusAkeneoPlugin\Behat\Context\Setup\AkeneoContext">
@@ -66,8 +64,10 @@
             <argument type="service" id="webgriffe_sylius_akeneo.behat.page.admin.queue_item.index" />
         </service>
 
-        <service id="webgriffe_sylius_akeneo.behat.context.ui.admin.managing_products" class="Tests\Webgriffe\SyliusAkeneoPlugin\Behat\Context\Ui\Admin\ManagingProducts">
+        <service id="webgriffe_sylius_akeneo.behat.context.ui.admin.managing_products" class="Tests\Webgriffe\SyliusAkeneoPlugin\Behat\Context\Ui\Admin\ManagingProductsContext">
             <argument type="service" id="sylius.behat.page.admin.product.index" />
+            <argument type="service" id="sylius.behat.java_script_test_helper" />
+            <argument type="service" id="sylius.behat.notification_checker" />
         </service>
 
         <service id="webgriffe_sylius_akeneo.behat.page.admin.queue_item.index" parent="sylius.behat.page.admin.crud.index" class="Tests\Webgriffe\SyliusAkeneoPlugin\Behat\Page\Admin\QueueItem\IndexPage" public="false">
